Add master and worker timers

+++ compss/runtime/adaptors/commons/src/main/java/es/bsc/compss/types/execution/Invocation.java
@@ -50,6 +50,6 @@ public interface Invocation {
 
     public List<String> getSlaveNodesNames();
 
-    public int getTimeOut();
+    public long getTimeOut();
 
 }
