+++ compss/runtime/adaptors/gat/worker/src/main/java/es/bsc/compss/gat/worker/implementations/COMPSsDefinition.java
@@ -82,8 +82,8 @@ public class COMPSsDefinition extends ImplementationDefinition {
     }
 
     @Override
-    public int getTimeOut() {
-        return 0;
+    public long getTimeOut() {
+        return 0L;
     }
 
 }
