+++ compss/runtime/adaptors/gat/worker/src/main/java/es/bsc/compss/gat/worker/implementations/DecafDefinition.java
@@ -82,8 +82,8 @@ public class DecafDefinition extends ImplementationDefinition {
     }
 
     @Override
-    public int getTimeOut() {
-        return 0;
+    public long getTimeOut() {
+        return 0L;
     }
 
 }
