+++ compss/runtime/adaptors/gat/worker/src/main/java/es/bsc/compss/gat/worker/implementations/OpenCLDefinition.java
@@ -66,8 +66,8 @@ public class OpenCLDefinition extends ImplementationDefinition {
     }
 
     @Override
-    public int getTimeOut() {
-        return 0;
+    public long getTimeOut() {
+        return 0L;
     }
 
 }
